feat: add unittests for core logic (#51)
* feat: add tests

* add tests for settings

* finish up, ready to merge

* cargo update

+++ src/db.rs
@@ -5,7 +5,8 @@ pub use sqlx::{query, query_as, Error, FromRow, Row};
 pub async fn connect(conn: &str, max_conn: Option<u32>) -> Result<SqlitePool, sqlx::Error> {
     let max_conn = max_conn.unwrap_or(5);
 
-    if !std::path::Path::new(conn).exists() {
+    if !conn.contains(":memory:") && !std::path::Path::new(conn).exists() {
+        std::fs::create_dir_all(std::path::Path::new(conn).parent().unwrap())?;
         std::fs::File::create(conn)?;
     }
